CSP: Block nonces for elements with duplicate attributes.
We throw away duplicate attributes when converting an `HTMLToken`
into an `AtomicHTMLToken` for use in `HTMLConstructionSite` (this
matches the specced behavior at [1]). That, unfortunately, means
that the checks at [2] are more or less impotent.

This patch adds a flag to both `HTMLScriptElement` and
`SVGScriptElement` that tracks whether or not the element was
constructed with duplicate attributes, and treats those elements as
non-nonceable. It also adds some counters; perhaps it's possible to
be more aggressive about the existing behavior for repeated attributes,
as it surprised me.

[1]: https://html.spec.whatwg.org/#attribute-name-state
[2]: https://w3c.github.io/webappsec-csp/#is-element-nonceable

Bug: 740615
Change-Id: I515753885167ac4acd7d0008e6aa81ce9f6f61cf
Reviewed-on: https://chromium-review.googlesource.com/566822
Commit-Queue: Mike West <mkwst@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#486752}
WPT-Export-Revision: 2546fab4295486a3adac0273e4a2eb61ad2a815b